# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Headless E-commerce API",
    "version": "1.0.0",
    "description": "A comprehensive headless e-commerce backend API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/cart": {
      "post": {
        "summary": "Create or get cart",
        "tags": [
          "Cart"
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Existing cart token (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cart created or retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cart/{token}": {
      "get": {
        "summary": "Get cart by token",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cart not found"
          }
        }
      }
    },
    "/api/cart/{token}/items": {
      "post": {
        "summary": "Add item to cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "productId",
                  "variantId",
                  "quantity"
                ],
                "properties": {
                  "productId": {
                    "type": "string"
                  },
                  "variantId": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added to cart",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or insufficient stock"
          },
          "404": {
            "description": "Cart or product not found"
          }
        }
      }
    },
    "/api/cart/{token}/items/{productId}/{variantId}": {
      "put": {
        "summary": "Update item quantity in cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          },
          {
            "in": "path",
            "name": "variantId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Variant ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "quantity"
                ],
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item quantity updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid quantity or insufficient stock"
          },
          "404": {
            "description": "Cart or item not found"
          }
        }
      },
      "delete": {
        "summary": "Remove item from cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          },
          {
            "in": "path",
            "name": "productId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Product ID"
          },
          {
            "in": "path",
            "name": "variantId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Variant ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Item removed from cart",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cart or item not found"
          }
        }
      }
    },
    "/api/cart/{token}/clear": {
      "post": {
        "summary": "Clear cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          }
        ],
        "responses": {
          "200": {
            "description": "Cart cleared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cart not found"
          }
        }
      }
    },
    "/api/cart/{token}/promo": {
      "post": {
        "summary": "Apply promo code to cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code"
                ],
                "properties": {
                  "code": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Promo code applied",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired promo code"
          },
          "404": {
            "description": "Cart not found"
          }
        }
      },
      "delete": {
        "summary": "Remove promo code from cart",
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Cart token"
          }
        ],
        "responses": {
          "200": {
            "description": "Promo code removed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Cart"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Cart not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cart": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Cart ID"
          },
          "token": {
            "type": "string",
            "description": "Cart token"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "quantity": {
                  "type": "integer"
                },
                "subtotal": {
                  "type": "number"
                }
              }
            }
          },
          "subtotal": {
            "type": "number",
            "description": "Cart subtotal"
          },
          "discount": {
            "type": "number",
            "description": "Applied discount amount"
          },
          "total": {
            "type": "number",
            "description": "Cart total after discount"
          },
          "promoCode": {
            "type": "string",
            "description": "Applied promo code"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "Cart expiration date"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}